blog page
In this project; There is a registration page and a login page. After the login page, we can access our own page, create a new post, delete or update the post we created. We do not have the opportunity to update or delete a post created by another user. I created my own database with the account I opened from the mongodb atlas.